requirements.txt
fastapi~=0.72.0
uvicorn
cassandra-driver~=3.25.0
python-dotenv
email-validator
argon2-cffi
pytest


pydantic~=1.9.0